requirements.txt
torch
accelerate
easydict
torchvision
numpy
tqdm
PyYaml
tensorboard
pandas